accessibility fixes and renaming for typo
